requirements.txt
aiohttp==3.8.1
linkpreview==0.2.0
nextcord==2.2.0
python-dotenv==0.21.0
tweepy==4.10.1
